Add auxiliary methods to CSP

+++ src/base/coloring.h
@@ -3,5 +3,5 @@
 #include <inttypes.h>
 #include <vector>
 
-using Color = uint8_t;
+using Color = size_t;
 using Coloring = std::vector<Color>;
